+++ changelog.md
@@ -6,6 +6,11 @@ next
   `getGCStats` and `applyGCStats`, which have also been deprecated and will be
   removed in the next major `criterion` release.
 
+* Export `Criterion.Main.Options.config`.
+
+* Export `toBenchmarkable`, which behaves like the `Benchmarkable` constructor
+  did prior to `criterion-1.2.0.0`.
+
 1.2.0.0
 
 * Use `statistics-0.14`.